+++ src/Symfony/Component/Messenger/Bridge/Doctrine/Tests/Transport/ConnectionTest.php
@@ -179,7 +179,7 @@ class ConnectionTest extends TestCase
     /**
      * @dataProvider buildConfigurationProvider
      */
-    public function testBuildConfiguration(string $dsn, array $options, string $expectedConnection, string $expectedTableName, int $expectedRedeliverTimeout, string $expectedQueue, bool $expectedAutoSetup)
+    public function testBuildConfiguration(string $dsn, array $options, string $expectedConnection, string $expectedTableName, int $expectedRedeliverTimeout, string $expectedQueue, bool $expectedAutoSetup, bool $expectedBinaryBody)
     {
         $config = Connection::buildConfiguration($dsn, $options);
         $this->assertEquals($expectedConnection, $config['connection']);
@@ -187,6 +187,7 @@ class ConnectionTest extends TestCase
         $this->assertEquals($expectedRedeliverTimeout, $config['redeliver_timeout']);
         $this->assertEquals($expectedQueue, $config['queue_name']);
         $this->assertEquals($expectedAutoSetup, $config['auto_setup']);
+        $this->assertEquals($expectedBinaryBody, $config['binary_body']);
     }
 
     public static function buildConfigurationProvider(): iterable
@@ -199,6 +200,7 @@ class ConnectionTest extends TestCase
             'expectedRedeliverTimeout' => 3600,
             'expectedQueue' => 'default',
             'expectedAutoSetup' => true,
+            'expectedBinaryBody' => false,
         ];
 
         yield 'test options array' => [
@@ -214,6 +216,7 @@ class ConnectionTest extends TestCase
             'expectedRedeliverTimeout' => 1800,
             'expectedQueue' => 'important',
             'expectedAutoSetup' => false,
+            'expectedBinaryBody' => false,
         ];
 
         yield 'options from dsn' => [
@@ -224,6 +227,7 @@ class ConnectionTest extends TestCase
             'expectedRedeliverTimeout' => 1200,
             'expectedQueue' => 'normal',
             'expectedAutoSetup' => false,
+            'expectedBinaryBody' => false,
         ];
 
         yield 'options from dsn array wins over options from options' => [
@@ -239,6 +243,7 @@ class ConnectionTest extends TestCase
             'expectedRedeliverTimeout' => 1200,
             'expectedQueue' => 'normal',
             'expectedAutoSetup' => true,
+            'expectedBinaryBody' => false,
         ];
 
         yield 'options from dsn with falsey boolean' => [
@@ -249,6 +254,7 @@ class ConnectionTest extends TestCase
             'expectedRedeliverTimeout' => 3600,
             'expectedQueue' => 'default',
             'expectedAutoSetup' => false,
+            'expectedBinaryBody' => false,
         ];
 
         yield 'options from dsn with thruthy boolean' => [
@@ -259,6 +265,7 @@ class ConnectionTest extends TestCase
             'expectedRedeliverTimeout' => 3600,
             'expectedQueue' => 'default',
             'expectedAutoSetup' => true,
+            'expectedBinaryBody' => false,
         ];
     }